Add in void display_board function
This function is responsible for displaying the game board on the screen. It provides a visual representation of the current state of the board to the users, assisting each player in deciding where to place their move.

This change can be summarized as follows:

Addition of the display_board function.
This function prints the state of the game board on the console.
It shows the current status of each player and the possible moves to the users.