Add basic Makefile for reversi project
adding a simple Makefile for the Reversi project. This Makefile will help to automate the build process and simplify the steps required to compile and link the project.

Here are the details:

Creation of a Makefile for the Reversi project.
It includes targets for compiling the main.c and reversi.c files into object files.
It also includes a target for linking these object files into an executable named "reversi".
A 'clean' target has also been added to clean up the project by removing all object files and the executable.